diy-streamlit-app/requirements.txt
streamlit
pandas
minsearch
requests
Pillow
